Round off final result rather than truncate for decimal outputs
Based on the number of shown digits parameter, the displayed result is rounded off
to the right number of decimal places.

If the value lies between (-1, 1), then the value is directly rounded off.

If the value is otherwise, then the length integer part of the result is calculated
and the result is rounded to the remaining decimal places.

If the length of the integer part of the result is greater than the number of digits
shown parameter, the result is displayed unchanged/unmodified.

eg: With number of shown digits set to 6, calculation sin(45) yields 0.707106
eg: With number of shown digits set to 15, calculation sin(45)*100 yields 70.7106781187. This is because
the length of the result calculated itself is 12.
eg: With number of shown digits set to 6, calculation sin(45)*100 yields 70.7107
eg: With number of shown digits set to 6, calculation 123.456+0.0007 yields 123.457

How it used to work before the patch
For non-trignometric decimal calculations, the display result was truncated to the number of
shown digits. No rounding off used to take place. For trignometric decimal calculations, the
result is once again truncated without rounding off.

eg: With number of shown digits set to 9, cos(90) yields 6.12323399x10^-17
eg: With number of shown digits set to 3, cos(90) yields 6.12x10^-17
eg: With number of shown digits set to 6, sin(45) yields 0.707106
eg: With number of shown digits set to 15, sin(45)*100 yields 70.7106781186547
eg: With number of shown digits set to 15, sin(45) yields 0.707106781186547
eg: With number of shown digits set to 6, calculation 123.456+0.0007 yields 123.456
eg: With number of shown digits set to 6, calculation sin(45)*100 yields 70.7106