Prevent Gtk-CRITICAL warning when removing toolbar items
Issue:

When the window gets resized the following warning occurs, `(sugar-activity3:5548): Gtk-CRITICAL **: 07:59:30.659: gtk_toolbar_remove: assertion 'content_to_remove != NULL' failed`

Change:

Added a safety check to ensure that a toolbar item has toolbar as a parent before attempting to remove it. This resolves a Gtk-CRITICAL warning that occurred when trying to remove an item that was not in the toolbar.